Start of refactoring aroud some of the entry level flow, handling of command line options vs whats in the config file now that there is one
